allow create or assert existing cart rule. Requires restoring only those scenarios that includes cart rules without coes (that affects all other scenarios)
