+++ tests/Integration/Behaviour/Features/Context/Domain/CartRule/CartRuleAssertionFeatureContext.php
@@ -30,20 +30,17 @@ namespace Tests\Integration\Behaviour\Features\Context\Domain\CartRule;
 
 use Behat\Gherkin\Node\TableNode;
 use PHPUnit\Framework\Assert;
-use PrestaShop\Decimal\DecimalNumber;
 use PrestaShop\PrestaShop\Core\Domain\CartRule\Exception\CartRuleConstraintException;
 use PrestaShop\PrestaShop\Core\Domain\CartRule\Exception\CartRuleNotFoundException;
 use PrestaShop\PrestaShop\Core\Domain\CartRule\Query\GetCartRuleForEditing;
 use PrestaShop\PrestaShop\Core\Domain\CartRule\QueryResult\CartRuleForEditing;
-use PrestaShop\PrestaShop\Core\Util\DateTime\DateTime;
 use RuntimeException;
 use Tests\Integration\Behaviour\Features\Context\Util\NoExceptionAlthoughExpectedException;
-use Tests\Integration\Behaviour\Features\Context\Util\PrimitiveUtils;
 
 class CartRuleAssertionFeatureContext extends AbstractCartRuleFeatureContext
 {
     /**
-     * @Then Cart rule with reference :cartRuleReference is enabled
+     * @Then cart rule with reference :cartRuleReference is enabled
      *
      * @param string $cartRuleReference
      *
@@ -62,7 +59,7 @@ class CartRuleAssertionFeatureContext extends AbstractCartRuleFeatureContext
     }
 
     /**
-     * @Then Cart rule with reference :cartRuleReference is disabled
+     * @Then cart rule with reference :cartRuleReference is disabled
      *
      * @param string $cartRuleReference
      *
@@ -145,200 +142,11 @@ class CartRuleAssertionFeatureContext extends AbstractCartRuleFeatureContext
      * @param string $cartRuleReference
      * @param TableNode $tableNode
      */
-    public function assertCartRuleProperties(string $cartRuleReference, TableNode $tableNode): void
+    public function assertCartRule(string $cartRuleReference, TableNode $tableNode): void
     {
-        /** @var CartRuleForEditing $editableCartRule */
-        $editableCartRule = $this->getQueryBus()->handle(
-            new GetCartRuleForEditing($this->getSharedStorage()->get($cartRuleReference))
+        $this->assertCartRuleProperties(
+            $this->getCartRuleForEditing($this->getSharedStorage()->get($cartRuleReference)),
+            $this->localizeByColumns($tableNode)
         );
-
-        $data = $this->localizeByRows($tableNode);
-        $information = $editableCartRule->getInformation();
-        $conditions = $editableCartRule->getConditions();
-        $actions = $editableCartRule->getActions();
-
-        if (isset($data['name'])) {
-            Assert::assertSame($data['name'], $information->getLocalizedNames(), 'Unexpected localized name');
-        }
-        if (isset($data['description'])) {
-            Assert::assertSame($data['description'], $information->getDescription(), 'Unexpected description');
-        }
-        if (isset($data['highlight'])) {
-            Assert::assertSame(
-                PrimitiveUtils::castStringBooleanIntoBoolean($data['highlight']),
-                $information->isHighlight(),
-                'Unexpected highlight'
-            );
-        }
-        if (isset($data['allow_partial_use'])) {
-            Assert::assertSame(
-                PrimitiveUtils::castStringBooleanIntoBoolean($data['allow_partial_use']),
-                $information->isPartialUse(),
-                'Unexpected partial use'
-            );
-        }
-        if (isset($data['active'])) {
-            Assert::assertSame(
-                PrimitiveUtils::castStringBooleanIntoBoolean($data['active']),
-                $information->isEnabled(),
-                'Unexpected active property'
-            );
-        }
-        if (isset($data['code'])) {
-            Assert::assertSame($data['code'], $information->getCode(), 'Unexpected code');
-        }
-        if (isset($data['customer'])) {
-            Assert::assertSame(
-                !empty($data['customer']) ? $this->getSharedStorage()->get($data['customer']) : 0,
-                $conditions->getCustomerId()->getValue(),
-                'Unexpected customer id'
-            );
-        }
-        if (isset($data['priority'])) {
-            Assert::assertSame((int) $data['priority'], $information->getPriority(), 'Unexpected priority');
-        }
-        if (isset($data['valid_from'])) {
-            Assert::assertEquals(
-                $data['valid_from'],
-                $conditions->getDateFrom()->format(DateTime::DEFAULT_DATETIME_FORMAT),
-                'Unexpected valid_from'
-            );
-        }
-        if (isset($data['valid_to'])) {
-            Assert::assertEquals(
-                $data['valid_to'],
-                $conditions->getDateTo()->format(DateTime::DEFAULT_DATETIME_FORMAT),
-                'Unexpected valid_to'
-            );
-        }
-        if (isset($data['total quantity'])) {
-            Assert::assertSame((int) $data['total quantity'], $conditions->getQuantity(), 'Unexpected quantity');
-        }
-        if (isset($data['quantity per user'])) {
-            Assert::assertSame((int) $data['quantity per user'], $conditions->getQuantityPerUser(), 'Unexpected quantity per user');
-        }
-
-        if (isset($data['minimum_amount'])) {
-            if (empty($data['minimum_amount'])) {
-                Assert::assertNull($conditions->getMinimum(), 'unexpected minimum_amount');
-            } else {
-                $minimum = $conditions->getMinimum();
-                Assert::assertTrue(
-                    $minimum->getAmount()->equals(new DecimalNumber($data['minimum_amount'])),
-                    'Unexpected minimum_amount'
-                );
-
-                Assert::assertSame(
-                    PrimitiveUtils::castStringBooleanIntoBoolean($data['minimum_amount_tax_included']),
-                    $minimum->isAmountTax(),
-                    'Unexpected minimum_amount_tax_included'
-                );
-
-                Assert::assertSame(
-                    $this->getSharedStorage()->get($data['minimum_amount_currency']),
-                    $minimum->getCurrencyId(),
-                    'Unexpected minimum_amount_currency'
-                );
-
-                Assert::assertSame(
-                    PrimitiveUtils::castStringBooleanIntoBoolean($data['minimum_amount_shipping_included']),
-                    $minimum->isShipping(),
-                    'Unexpected minimum_amount_shipping_included'
-                );
-            }
-        }
-
-        if (isset($data['discount_application_type'])) {
-            $expectedDiscountApplicationType = $data['discount_application_type'];
-            Assert::assertSame(
-                $expectedDiscountApplicationType,
-                $actions->getDiscountApplicationType(),
-                'Unexpected discount_application_type'
-            );
-
-            Assert::assertSame(
-                !empty($data['discount_product']) ? $this->getSharedStorage()->get($data['discount_product']) : null,
-                $actions->getReduction()->getProductId(),
-                'Unexpected discount_product'
-            );
-        }
-
-        if (isset($data['discount_percentage'])) {
-            Assert::assertTrue(
-                $actions->getReduction()->getPercent()->equals(new DecimalNumber($data['discount_percentage'])),
-                'Unexpected discount_percentage'
-            );
-        }
-
-        if (isset($data['discount_amount'])) {
-            Assert::assertTrue(
-                $actions->getReduction()->getAmount()->equals(new DecimalNumber($data['discount_amount'])),
-                'Unexpected discount_amount'
-            );
-        }
-
-        if (isset($data['discount_currency'])) {
-            if ($data['discount_currency'] === '') {
-                Assert::assertNull(
-                    $actions->getReduction()->getCurrencyId(),
-                    'Unexpected discount_currency'
-                );
-            } else {
-                Assert::assertSame(
-                    $this->getSharedStorage()->get($data['discount_currency']),
-                    $actions->getReduction()->getCurrencyId(),
-                    'Unexpected discount_currency'
-                );
-            }
-        }
-
-        if (isset($data['discount_includes_tax'])) {
-            Assert::assertSame(
-                PrimitiveUtils::castStringBooleanIntoBoolean($data['discount_includes_tax']),
-                $actions->getReduction()->isTax(),
-                'Unexpected discount_includes_tax'
-            );
-        }
-
-        if (isset($data['apply_to_discounted_products'])) {
-            Assert::assertSame(
-                PrimitiveUtils::castStringBooleanIntoBoolean($data['apply_to_discounted_products']),
-                $actions->getReduction()->applyToDiscountedProducts(),
-                'Unexpected apply_to_discounted_products'
-            );
-        }
-
-        if (isset($data['free_shipping'])) {
-            Assert::assertSame(
-                PrimitiveUtils::castStringBooleanIntoBoolean($data['free_shipping']),
-                $actions->isFreeShipping(),
-                'Unexpected free_shipping'
-            );
-        }
-
-        if (isset($data['gift_product'])) {
-            Assert::assertSame(
-                $this->getSharedStorage()->get($data['gift_product']),
-                $actions->getGiftProductId(),
-                'Unexpected gift_product'
-            );
-        }
-
-        if (isset($data['gift_combination'])) {
-            Assert::assertSame(
-                !empty($data['gift_combination']) ? $this->getSharedStorage()->get($data['gift_combination']) : null,
-                $actions->getGiftCombinationId(),
-                'Unexpected gift_combination'
-            );
-        }
-
-        if (isset($data['restricted cart rules'])) {
-            $expectedRestrictedCartRuleIds = $this->referencesToIds($data['restricted cart rules']);
-            Assert::assertSame(
-                $expectedRestrictedCartRuleIds,
-                $editableCartRule->getConditions()->getRestrictions()->restrictedCartRuleIds,
-                'Unexpected cart rule restrictions'
-            );
-        }
     }
 }
