+++ tests/Integration/Behaviour/Features/Scenario/Cart/CartRule/FrontOffice/cart_rule_validation.feature
@@ -9,7 +9,7 @@ Feature: Cart rule application is validated before it is applied to cart
     Given there is customer "testCustomer" with email "pub@prestashop.com"
     And country "US" is enabled
     And there is a currency named "usd" with iso code "USD" and exchange rate of 0.92
-    And I create cart rule "cart_rule_1" with following properties:
+    And there is a cart rule "cart_rule_1" with following properties:
       | name[en-US]                  | cartrule1              |
       | total_quantity               | 1000                   |
       | quantity_per_user            | 1000                   |
@@ -18,7 +18,7 @@ Feature: Cart rule application is validated before it is applied to cart
       | discount_percentage          | 50                     |
       | apply_to_discounted_products | true                   |
       | discount_application_type    | order_without_shipping |
-    And I create cart rule "cart_rule_2" with following properties:
+    And there is a cart rule "cart_rule_2" with following properties:
       | name[en-US]                  | cartrule2              |
       | total_quantity               | 10                     |
       | quantity_per_user            | 10                     |
@@ -29,7 +29,7 @@ Feature: Cart rule application is validated before it is applied to cart
       | discount_includes_tax        | false                  |
       | apply_to_discounted_products | true                   |
       | discount_application_type    | order_without_shipping |
-    And I create cart rule "cart_rule_3" with following properties:
+    And there is a cart rule "cart_rule_3" with following properties:
       | name[en-US]       | cartrule3 |
       | total_quantity    | 10        |
       | quantity_per_user | 10        |
@@ -67,7 +67,6 @@ Feature: Cart rule application is validated before it is applied to cart
     And total cart shipping fees should be 0.0 tax excluded
     And my cart total should be 0.0 tax excluded
 
-  @restore-cart-rules-before-scenario
   Scenario: Cart rule cannot be applied again when it is already in the cart
     Given I have an empty default cart
     And shop configuration for "PS_CART_RULE_FEATURE_ACTIVE" is set to 1
